renamed component to fix spelling
